# Improves coordinate conversion by caching last converter