- Problem: To find a converter in the std::unordered_map is more expensive than single conversion (~factor 2)
- Keeps a shared_ptr of the last used converters s.t. this cached converter hits for all consecutively converted conversions from the same system
- thread_local cache, no need for mutexes