@stmitt 2025-03-25
Nice, benchmarked with the testcases of #708 

before: 
```
VectorTileGeometryHandler
-------------------------------------------------------------------------------
/Users/stefan/Developer/maps-core/shared/test/TestGeometryHandler.cpp:77
...............................................................................

benchmark name                       samples       iterations    est run time
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Benchmark regular                              100             1     8.13326 s 
                                        82.1057 ms    81.8174 ms    82.9102 ms 
                                        2.25246 ms    1.00073 ms    4.81748 ms 
                                                                               
Benchmark relief                               100             1     1.01462 m 
                                        602.386 ms    599.007 ms    608.155 ms 
                                        22.0472 ms    15.1811 ms    37.7165 ms 
```


after:
```
VectorTileGeometryHandler
-------------------------------------------------------------------------------
/Users/stefan/Developer/maps-core/shared/test/TestGeometryHandler.cpp:77
...............................................................................

benchmark name                       samples       iterations    est run time
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Benchmark regular                              100             1     8.09137 s 
                                         81.106 ms    80.8128 ms    81.6456 ms 
                                        1.97724 ms    1.21368 ms     3.1226 ms 
                                                                               
Benchmark relief                               100             1     55.4872 s 
                                        556.549 ms    555.388 ms    558.259 ms 
                                        7.11342 ms    5.23602 ms     9.5965 ms 
```